Refactored URLs to be more beautiful. Fixes #52.
Also moved login related stuff out of ApplicationController.
