+++ app/cobase/play/user/ApplicationController.scala
@@ -4,7 +4,7 @@ import javax.inject.Inject
 
 import cobase.post.PostService
 import cobase.user.{GroupService, User}
-import com.mohiva.play.silhouette.api.{Environment, LogoutEvent, Silhouette}
+import com.mohiva.play.silhouette.api.{Environment, Silhouette}
 import com.mohiva.play.silhouette.impl.authenticators.SessionAuthenticator
 
 import scala.concurrent.Future
@@ -31,39 +31,4 @@ class ApplicationController @Inject() (implicit val env: Environment[User, Sessi
     Future.successful(Ok(views.html.home(request.identity, groupLinks, dashboardPosts)))
   }
 
-  /**
-   * Handles the Sign In action.
-   *
-   * @return The result to display.
-   */
-  def signIn = UserAwareAction.async { implicit request =>
-    request.identity match {
-      case Some(user) => Future.successful(Redirect(routes.ApplicationController.index))
-      case None => Future.successful(Ok(views.html.signIn(SignInForm.form)))
-    }
-  }
-
-  /**
-   * Handles the Sign Up action.
-   *
-   * @return The result to display.
-   */
-  def signUp = UserAwareAction.async { implicit request =>
-    request.identity match {
-      case Some(user) => Future.successful(Redirect(routes.ApplicationController.index))
-      case None => Future.successful(Ok(views.html.signUp(SignUpForm.form)))
-    }
-  }
-
-  /**
-   * Handles the Sign Out action.
-   *
-   * @return The result to display.
-   */
-  def signOut = SecuredAction.async { implicit request =>
-    val result = Future.successful(Redirect(routes.ApplicationController.index))
-    env.eventBus.publish(LogoutEvent(request.identity, request, request2lang))
-
-    request.authenticator.discard(result)
-  }
 }
