+++ app/cobase/play/user/CredentialsAuthController.scala
@@ -48,7 +48,7 @@ class CredentialsAuthController @Inject() (
         }
       }.recover {
         case e: ProviderException =>
-          Redirect(routes.ApplicationController.signIn()).flashing("error" -> Messages("invalid.credentials"))
+          Redirect(routes.AuthenticationController.signIn).flashing("error" -> Messages("invalid.credentials"))
       }
     )
   }
