+++ app/cobase/play/user/SocialAuthController.scala
@@ -54,7 +54,7 @@ class SocialAuthController @Inject() (
     }).recover {
       case e: ProviderException =>
         logger.error("Unexpected provider error", e)
-        Redirect(routes.ApplicationController.signIn()).flashing("error" -> Messages("could.not.authenticate"))
+        Redirect(routes.AuthenticationController.signIn).flashing("error" -> Messages("could.not.authenticate"))
     }
   }
 }
