+++ app/views/editGroup.scala.html
@@ -25,7 +25,7 @@
                 @request.flash.get("error").map { msg =>
                 <p class="register-box-msg"><strong>Error:</strong> @msg</p>
                 }
-                @helper.form(action = groupControllerRoutes.updateGroup(group.id)) {
+                @helper.form(action = groupControllerRoutes.editGroup(group.id)) {
                 @text(groupForm("title"), "Title", icon = "")
                 @tags(groupForm("tags"), "Twitter hashtags")
                 <div class="row">
