+++ app/views/editPost.scala.html
@@ -27,7 +27,7 @@
                 @request.flash.get("error").map { msg =>
                 <p class="register-box-msg"><strong>Error:</strong> @msg</p>
                 }
-                @helper.form(action = postControllerRoutes.updatePost(post.id)) {
+                @helper.form(action = postControllerRoutes.editPost(post.id)) {
                 @textarea(postForm("content"), "Post content", icon = "")
                 <div class="row">
                     <div class="col-xs-4">
