+++ app/views/group.scala.html
@@ -43,7 +43,7 @@
                                 <h3 class="box-title">What's on your mind?</h3>
                             </div>
                             <div class="box-body">
-                                @helper.form(action = postControllerRoutes.createPost(group.get.id)) {
+                                @helper.form(action = postControllerRoutes.addPost(group.get.id)) {
                                 @textarea(postForm("content"), "Post content", icon = "")
                                 <div class="row">
                                     <div class="col-xs-4">
