+++ app/views/newGroup.scala.html
@@ -24,7 +24,7 @@
                     @request.flash.get("error").map { msg =>
                         <p class="register-box-msg"><strong>Error:</strong> @msg</p>
                     }
-                    @helper.form(action = groupControllerRoutes.createGroup()) {
+                    @helper.form(action = groupControllerRoutes.addGroup()) {
                         @text(groupForm("title"), "Title", icon = "")
                         @tags(groupForm("tags"), "Twitter hashtags")
                         <div class="row">
