+++ app/views/signIn.scala.html
@@ -4,7 +4,7 @@
 @import views.html.helpers._
 
 @import cobase.user.User
-@import cobase.play.user.routes.{CredentialsAuthController => credentialsAuthControllerRoutes, ApplicationController => applicationControllerRoutes}
+@import cobase.play.user.routes.{CredentialsAuthController => credentialsAuthControllerRoutes, SignUpController => signUpControllerRoutes}
 
 @(signInForm: Form[Credentials])(implicit env: Environment[User, _], request: RequestHeader)
 
@@ -34,7 +34,7 @@
                     </div><!-- /.col -->
                 </div>
             }
-            <p class="not-a-member top20">Not a member? <a href="@applicationControllerRoutes.signUp">Sign up now</a></p>
+            <p class="not-a-member top20">Not a member? <a href="@signUpControllerRoutes.signUp">Sign up now</a></p>
         </div>
     </div>
 }
